+++ src/com/intellij/idea/plugin/hybris/project/configurators/MavenConfigurator.kt
@@ -1,7 +1,7 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -45,11 +45,12 @@ class MavenConfigurator {
             .map { File(it, MavenConstants.POM_XML) }
             .filter { it.exists() && it.isFile }
             .mapNotNull { VfsUtil.findFileByIoFile(it, true) }
-            .map {
-                val builder = MavenProjectBuilder()
-                builder.isUpdate = MavenProjectsManager.getInstance(project).isMavenizedProject
-                builder.setFileToImport(it)
-                builder
+            .map { vf ->
+                MavenProjectBuilder()
+                    .also {
+                        it.setUpdate(MavenProjectsManager.getInstance(project).isMavenizedProject)
+                        it.setFileToImport(vf)
+                    }
             }
             .filter {
                 val path = it.rootPath